// src/app/shared/models/parametro.model.ts
export class Parametro {
    idParametro?:Number;
    codigo?:string;
    valorPrimario?: string;
    valorSecundario?:string;
    idTipoParametro?: Number;
    idParametroPadre?: Number;
    valorTerciario?: string;
    prefijo?: string;
    nombre?: string;
    descripcion?: string;
    estado?: string
}